meta: update atime in inode get operation
Update atime in inode get operation, so that temporary file validation
can be extended.

A temporary file is defined as file which is created and removed
immediately by a process, and only the one who is holding the file
descriptor can manipulate the file. However, metanode has difficulties
telling whether a zero-nlink inode is orphan or some user is still holding
the fd. In order to eliminate dirty orphan inode, there is a mechanism to
evict expired zero-nlink inode.

This patch avoids the wrong eviction of read-only temporary files.

Signed-off-by: Shuoran Liu <shuoranliu@gmail.com>