fix: exclude data partition only when connection refused
When choosing data partition to create extent, client will mark a dp as
exception if it gets an error response so that next time the dp will not
be selected. However, sometimes this failure is temporal due to heavy
network workload, and should not be excluded in the next selection.

Signed-off-by: Shuoran Liu <shuoranliu@gmail.com>