fix: metanode mistakenly delete empty dir inode
If a directory is empty, the nlink of the dir inode is 2. And metanode
may mistakenly delete the inode upon receiving an iunlink request.

For example, if we are renaming a directory, the client will send ilink
and iunlink requests to the dir inode during the Rename process, and
this inode will be wrongly deleted after iunlink.

Signed-off-by: Shuoran Liu <shuoranliu@gmail.com>