fix: stream traverse process never gets triggered in some situation
When a user opens a file and writes few bytes to it without fsync
or close, the traverse process will never get triggered because the
written bytes cached in buffer are less than the threshold to be
flushed.

So if there is no writes for a period of time, traverse should
flush the buffer and update metadata.

Signed-off-by: Shuoran Liu <shuoranliu@gmail.com>