fix: wait for data sync in close syscall
The Linux *close* syscall invokes the fuse Flush method which is not
implemented before. So the data might still in client's buffer after a
file is closed. Although there is no data persistence requirement for
the *close* syscall, we'd better wait for data to be synced so that
other clients can see the latest data after the file is closed in this
client.

Signed-off-by: Shuoran Liu <shuoranliu@gmail.com>