Remove: doc code of rocksdb in vendor
Remove document project code of rocksdb in vendor to pass the "Mixed
content - romantic-mclean-d9b22e".
This check is used to check if the html file in the project code
contains non-https links. It found that the documentation website
code for rocksdb that exists in the vendor directory of the project
contains non-https links.

DETAILS:
  In vendor/rocksdb-5.9.2/docs/_includes/post.html:
    img http://graph.facebook.com/{{ author.fbid }}/picture/
  In vendor/rocksdb-5.9.2/docs/_includes/plugins/slideshow.html:
    script http://fb.me/react-with-addons-0.13.1.min.js

Signed-off-by: Mofei Zhang <mofei2816@gmail.com>