Fix: memory overflow risk in object node.
Each objectnode can operate all volumes in the entire cluster. The
objectnode will load the volume and initialize the related ExtentClient
object. ExtentClient uses a global object pool to optimize memory. This
causes the objectnode to repeatedly create a global object pool during
operation. There is a risk of memory overflow.
The object pool in the ExtentClient is adjusted to be initialized by the
init () method to solve the problem of repeated creation.

Signed-off-by: Mofei Zhang <mofei2816@gmail.com>