Enhancement: optional not including ETag
When the object storage HeadObject and GetObject have no etag
information in the extended attributes of the file, the ETag header is
not returned. This solves the problem of data verification failure when
reading files written through the file system interface using the AWS S3
SDK through the object storage interface to a certain extent.

Signed-off-by: Mofei Zhang <mofei2816@gmail.com>