Enhancement: compatible with older version s3 credential
Added compatibility with 1.5 version of credential information for
signature verification and policy check of s3 interface.

In version 1.5, each volume has its own access key, which is stored
in the view information center of the volume to which it belongs.

In version 2.0, user security and authorization are added. Each user
has their own access key, and users can use this access key to operate
all their own volumes.

Increase the compatibility with the old version authentication method
to reduce the impact of online version 1.5 cluster upgrade to 2.0 on
online services.

Signed-off-by: Mofei Zhang <mofei2816@gmail.com>