enhance: limit the rate of Forget requests
The kernel is sending Forget requests without waiting for the responses,
so it is not limited by the value of FUSE_DEFAULT_MAX_BACKGROUND. And
there is a chance that kernel sends tons of Forget requests in a very
short time which runs out of memory.

This commit limits the rate of dealing with Forget requests to optimize
memory consumption.

Signed-off-by: Shuoran Liu <shuoranliu@gmail.com>