Fix: multipart part inode release issue
The inode of the part in multipart upload is not consistent with the
serialization and deserialization logic used when submitting and applying
through raft when it is released.

This results in the inability to successfully clean up useless part inode
data structures after users have passed the CompleteMultipartUpload via
S3-compatible interface.

Signed-off-by: Mofei Zhang <mofei2816@gmail.com>