Fix: S3 response when volume do not exists
When initializing the volume, if the master explicitly responds that the
specified volume does not exist, it will not retry.

During the processing of the S3 request, both the signature and the
strategy are checked for the existence of the volume. NoSuchBucket
response is returned when the specified volume does not exist.

Signed-off-by: Mofei Zhang <mofei2816@gmail.com>