fix: mount failure for authorized user
A non-owner user can be authorized to mount with rw/ro permission.
However, the auth key is generated according to the user name which
leads to mount failure when getting volume info from master.

To reproduce:

1. create a volume named V with owner A.
2. authorize user B the rw permission to V.
3. mount V using user B with access and secret key.

The mount would fail.

This commit will not check owner auth key when using access key to
mount.

Signed-off-by: Shuoran Liu <shuoranliu@gmail.com>