Add definition for OpUpdateDataPartitionReplicas
In order to improve the efficiency of data repair after adding a new
replica, adjust the repair after the new replica is added to trigger the
Hong Kong style, and actively push the host information of the new
replicas to the members of the replication group after the master
updates the view.
This commit defines new command constants and related structures for the
process.
See'proto/admin_proto.go:234' and ‘proto/packet.go:124’ for details.

Signed-off-by: Mofei Zhang <mofei2816@gmail.com>