fix: change append extent key logic
Suppose there are two extent keys, [0,100) and [200,300), and we are
appending [150,250). The old logic will have 3 keys.

This commit changes the logic to keep only [0,100) and [150,250], to avoid
overlapping keys, but keeps the size to 300. This case happens only when
client extent handler sends partial data packets successfully and updates
metadata, then recover the remaining data packets to another extent
handler.

Signed-off-by: Shuoran Liu <shuoranliu@gmail.com>