Fix panic when datapartition removing current peer (self)
The reason for this problem is that the stop raft operation and the
raft apply member change log operation are concurrent, but the stop
raft operation did not wait for the runApply goroutine of raft to
finish before returning, which caused runApply to execute this
changemember log.
The raftstore has been set to nil.

Signed-off-by: Mofei Zhang <mofei2816@gmail.com>