fix: makes AppendExtentKeyWithCheck request idempotent
This commit makes AppendExtentKeyWithCheck request idempotent to avoid
conflict extents error when client retries.

For example,
    -> client writes [0, 100) to extent 25,
    -> send AppendExtentKeyWithCheck to metanode (offset: 0, size: 100, extentID:
25, discard extent: nil),
    -> client overwrites [0, 100) to extent 30,
    -> send AppendExtentKeyWithCheck to metanode (offset: 0, size:
100, extentID: 30, discard extent: 25),

Think about the following scenario: metanode managed to deal with the request,
but failed to respond to the client, and client retries with the same request.
Since AppendExtentKeyWithCheck is not idempotent, this will lead to
conflict extents error.

    -> client send AppendExtentKeyWithCheck to metanode (offset: 0,
size: 100, extentID: 30, discard extent: 25),
    -> when metanode is dealing with such request again, it expects no
discard extent, so it will respond a conflict extent error.

Reported-by: zcola <zhangrui90@163.com>
Signed-off-by: Shuoran Liu <shuoranliu@gmail.com>