metanode: Add new snapshot files
This patch add new snapshot files xxx_small to store metadata in a
new format. The file is composed of a header and multiple snapshot
blocks, which has fixed block size.

The header indicates the format version and the size of snapshot
block. Its size is 32 bytes and only the first 4 bytes are used
for now. Each block consits of continous entries.

If left space in one block is not enough to save one more entry,
the entry is saved to the next block. If one entry is too large
to save in one entire block, the entry is saved to the orginal
snapshot file.

In this format, the loaders could load different blocks
concurrently.

Signed-off-by: Sheng Yong <shengyong2021@gmail.com>