metanode: Store Inode to small snapshot file
If Inode is too large to be saved in one entire block, save it to
the orginal snapshot file (large snapshot file). Otherwise, save
Inode to small snapshot file.

If left space in one block is not enough to save one more Inode,
the Inode is saved to the next block.

As a result, two CRCs are returned.

Signed-off-by: Sheng Yong <shengyong2021@gmail.com>