metanode: Add loadCursor to save load status
When loading a snapshot file, we have to save current status so that
multiple goroutines could share load status, which contains the next
snapshot block to be loaded, the entries number already loaded, if
error occurs when loading.

To avoid too many lock operations, values are saved in per-goroutine
array or atomic variables. Helper methods are offered to handle
status.

Signed-off-by: Sheng Yong <shengyong2021@gmail.com>