fuse: Add HandleReadDirer implementation
This implements ReadDir for fuse.

In order to save current position of readdir, a map DirContexts is
introduced in `Dir struct'. Key Value pair (fuse.HandleID,
DirContext) is created and saved in DirContexts when readdir
starts, and is removed when parent directory is closed.

ReadDir reads a number of dirents from current position, thus a
DirContext, based on ReadDirLimit. After that, a new position
is saved in the same DirContext.

Signed-off-by: Sheng Yong <shengyong2021@gmail.com>