fuse: Make ReadDir compatible with old version metanode
If metanode is still running in old version, which has no
ReadDirLimit, ReadDir will always fail. So this patch checks if
ReadDirLimit is supported according to return value. If it is not
implemented, ReadDir will rollback to ReadDirAll and remember
this state in Super.

NOTE that, if metanode is in the latest version, it returns ENOSYS
if function is not implemented. However, if it in old version, it
may returns EIO for calling ReadDirLimited, so we also consider
EIO as function not implemented. This should be fixed in the feture.

Signed-off-by: Sheng Yong <shengyong2021@gmail.com>