change: append volume to client log path
When multi client mount on one machine, one client log will be
overwritten by another. To avoid this, we should separate the log path
of each client. To implement this, i attach volume name to the logDir in
the config file.

Besides, previous default logDir is root. I set the default logDir to
/var/log/chubaofs

Signed-off-by: wanglei469 <wanglei469@ke.com>