datanode: Repair broken tiny extents more aggressively
When repairing a broken tiny extent, only up to ~4G data is transmitted.
If a tiny extent is very large, it takes hours to repair the extent.

This patch tries to repair the broken tiny extent continously until
local extent size is equal to remote extent size.  And an atomic
`Reparing' state is added to ExtentInfo to avoid multiple repair-
goroutines repair the same extent.

Signed-off-by: Sheng Yong <shengyong2021@gmail.com>