master: Add /dataNode/delete & /metaNode/delete
APIs `/dataNode/update` and `/metaNode/update` are used to fix wrong
node id manually. They could change node id by adding a new node info
(address and id are combined as key) to rocksdb.

As a result, there exists more than one records of the same address
address in master's rocksdb. So we need new APIs to delete invalid
(or old) node information from rocksdb.

This patch addes `/dataNode/delete` and `/metaNode/delete`, whose
parameters are `addr` and `id`. If the key (addr + id) is the current
node which is still in use, it cannot be deleted.

Signed-off-by: Sheng Yong <shengyong2021@gmail.com>