fuse: Suspend saves dummy node for orphan dir inode
If a directory is removed, its metadata in metanode is cleaned up
immediately. After a while, its metadata in cfs-client's icache
also get expired if not evict in time. In this case, when we are
trying to suspend the cfs-client, cfs-client will try to get Attr
from metanode. However, metanode returns ENOENT. And suspend fails.
(cfs-client handling regular file in a different way.)

To avoid the failure, if such ENOENT occurs, we add a dummy Attr
which only records ModeDir.

Signed-off-by: Sheng Yong <shengyong2021@gmail.com>