cli: Fix failure of unmarshal Inode.Extents
When executing:
	cfs-cli compatibility meta ./ 192.168.0.22:17220 147

The following error returned:
```
[Meta partition is 147, verify result]
The number of dentry is 0, all dentry are consistent
Error: json: cannot unmarshal object into Go struct field Inode.Extents of type []proto.ExtentKey
```

The reason is that the definitions of `Extents' of `Inode' in cli and
metanode are different. The former is defined as a slice, while the
latter is defined as a non-exportable pointer. So ignore the key in
both marshal and unmarshal sides.

Signed-off-by: Sheng Yong <shengyong2021@gmail.com>