client-change: adjust timeout duration when pprof port is in use
The time for client to wait when port is in use is 1m. And the hint
is logged in the file. Now, the time changed to 10s and hint is
foreground to make the users know what the client is doing.
Besides, the log interval is also changed to approximately 5s to
make the screen clearer.

Signed-off-by: wanglei469 <wanglei469@ke.com>