Introducing auto unstable protection feature for Raft
Risk State is added to Raft to identify whether the current replication
group is in a stable or unstable state.

When the Sync configuration item of Raft Storage is set to false, once
it switches from a stable state to an unstable state, Raft will
immediately perform a Flush (Sync) operation on the Storage to ensure
that the existing data is reliably placed on the disk. And perform the
Sync operation after the Raft Log and HardState are persisted.

Switching between stable and unstable states:
1. The Leader determines whether all members are reliable through the
   follower heartbeat response time (within 3 ticks from the current
   time). If the conditions are not met, it will immediately switch
   itself from a stable state to an unstable state, and will notify
   other Follower members to switch to an unstable state through
   Heartbeat.

2. Follower will determine whether the leader state is reliable by
   detecting the time difference (within 3 ticks) from the last time
   it received the leader's heartbeat. If the conditions are not met,
   it will immediately switch itself to an unstable state.