Update: enhance snapshot on tiny extent and split senarios. etc...
.write on exist extent normally happened and create a new extent to write. but for snapshot scenarios extent be split
and the last extent to append write directly may occur conflict because the position may have the splited extent and effective,
thus we need make a additional jugement for GetEndForAppendW, if split happend then not reuse the extent.
update: add space log for datanode

.Snapshot mod append on tiny extent should use replay extent that may be new

.origin design use the end of old tiny extent id while modAppend, but that process not use get and put ways,
and it's not safe if parrelle. still can be hanged if parrellel already use the tiny extent id

.update tiny modAppend process from modAppend to tinyAppendWrite
.tiny write directly need with follower addrs or else repair later but read may not get it
.inode ek be truncated considered as split 2) code optimizition on createLowerVersion

.extent key be splited and be dropped through extDelCh and freelist, multi ext key belong to one extent
the eks before the last one should just the make a hole on the extent,but the last one should delete the whole extent.

Signed-off-by: leonrayang <changliang@oppo.com>