fix: mp replica attach and start have several steps, error happes may cause mp cann't offer service but raft alive
close: #2117

All master stopped, then restart one replica of the normal metapartion, once the meta replica start ,the raft group will became normal，However, due to the failed connection to the master node, the replica in the metanode may become abnormal, resulting in the service not working properly. If the replica becomes a leader, the service will return an error.

It's better stop raft while abnormal happend,but the raft stop already be set as todo, which may consider the some process on flight, so reorder the sequence temporily.

Signed-off-by: leonrayang <chl696@sina.com>