fix(client): During the process of performing append writes on a snapshot file, there is a possibility of incorrect updates to the extent cache.
When performing append writes after a snapshot, we attempt to reuse extents, resulting in a change in behavior between streaming and direct writes.
During this process, it is important to maintain proper isolation between them, especially when making incorrect modifications to the extent cache.
Such errors can lead to subsequent write errors.

Signed-off-by: leonrayang <chl696@sina.com>