@formatter:off
The base line of kernel client for RDMA.

It passes the basic test script of fio and dd.

chore(kernel): replacei state unconnected with disconnected.

The IBVSOCKETCONNSTATE_UNCONNECTED value 0 is confused.
So add one more state: IBVSOCKETCONNSTATE_DISCONNECTED.
If the datanode server is not ready. It happens to report disconnected.