@formatter:off fix(kernel): fix the bug id: 22397042
In recover function, the leader address maybe in the follower address list.
The leader is switched by reader or other action.
For write action, it causes datanode writes the same data block twice.
So the IO broken happens.

Fix method:
Use the dp->members.base[0] as the leader address always.