@formatter:off fix(kernel): restore the rdma log codes.
Root cause:
If two mount link use the same rdma link. It is confused to write logs.
After one link release the log buffer, how the second to update the log buffer?

Two method to fix it:
1. Compare the log buffer address. Only reuse the rdma link in the same mount link.
2. Remove the rdma log codes to client's log buffer.

The code restoring is the method 2.

Updates:
Delete the unused local parameters of extent.