@formatter:off fix(kernel): fix the bug id: 22507419
Issue:
The meta is lost in the metanode. So the read context is empty.
The download file is different from the original one.

Root cause:
Sometime, the recover writer just quit in the writer flush.
The tx rx inflight is 0. The ext_size is not updated yet.
The extent_stream_get_writer and cfs_extent_stream_flush maybe
called before extent_writer_recover upload.
The uploading is still ok. However the meta is not submit.